fastapi-tools/pyproject.toml
[project]
name = "nydev-fastapi-tools"
dynamic = ["version"]
readme = "readme.md"
license = {text = "MIT NON-AI License"}
authors = [
    {name = "Rajan Khullar", email = "rkhullar03@gmail.com"}
]
requires-python = ">= 3.10"
dependencies = [
    "httpx",
    "fastapi",
    "pydantic-settings"
]

[project.optional-dependencies]
server = ["uvicorn"]
lambda = ["mangum"]
mongodb = ["pymongo[aws,srv]"]

[project.urls]
Repository = "https://github.com/rkhullar/python-libraries/tree/main/fastapi-tools"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "fastapi_tools.__version__"}
